+++ src/code-generation/index.ts
@@ -1,2 +1,0 @@
-export { GoogleFirestoreRenderer } from './google-firestore-renderer.js';
-export { GoogleSpannerRenderer } from './google-spanner-renderer.js';