+++ src/functions/google-spanner/index.ts
@@ -1,4 +1,5 @@
 export { ModelSchemaExtractDatabaseForGoogleSpanner } from './extract-database.js';
+export { ModelGenerateTypeScriptDecoratorsForGoogleSpanner } from './generate-typescript-decorators.js';
 export { GoogleSpannerListDatabases } from './list-databases.js';
 export { DatabaseQueryRecordsForSpanner } from './query-records.js';
 export { GoogleSpannerWriteDatabases } from './write-databases.js';