+++ src/index.spec.ts
@@ -10,7 +10,6 @@ const heavyModules = [
   'google-auth-library',
   'google-gax',
   'googleapis',
-  'quicktype-core',
 ];
 
 describe('registerFunctions', () => {